% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{ecdna_dataframe_to_ampgenes}
\alias{ecdna_dataframe_to_ampgenes}
\title{Get list of genes on ecDNA}
\usage{
ecdna_dataframe_to_ampgenes(data)
}
\arguments{
\item{data}{data.frame from \code{\link[=read_ecdna]{read_ecdna()}}}
}
\value{
A long dataframe with 4 Cols
}
\description{
Get list of genes on ecDNA
}
\examples{
linx_dir = system.file(package="ecDNAutils", "CHP212")
data <- read_ecdna(linx_dir, "CHP212")
ecdna_dataframe_to_ampgenes(data)
}
