% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{get_ecDNA_chain_coordinates}
\alias{get_ecDNA_chain_coordinates}
\title{get_ecDNA_chain_coordinates}
\usage{
get_ecDNA_chain_coordinates(linx_dir, sample, cluster = NA, chainId = NA)
}
\arguments{
\item{linx_dir}{path to a directory containing linx files}

\item{sample}{name of sample to examine}

\item{cluster}{name of clusters of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be vector of elements, where element 1 of cluster corresponds to element 1 of chainId, element 2 of cluster corresponds to element 2 of chainId ... etc}

\item{chainId}{name of chain ids of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be vector of elements, where element 1 of chainId corresponds to element 1 of cluster, element 2 of chainId corresponds to element 2 of cluster ... etc}
}
\value{
coordinates of a single ecDNA molecule (data.frame)
}
\description{
Get coordinates of segments in ecDNA molecules
}
\examples{
linx_dir = system.file(package="utilitybeltlinx", "CHP212")
get_ecDNA_chain_coordinates(linx_dir, "CHP212", cluster=274, chainId = 8)
}
