% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{get_ecDNA_chain_coordinates_write_bed}
\alias{get_ecDNA_chain_coordinates_write_bed}
\title{ecDNA coords to bed}
\usage{
get_ecDNA_chain_coordinates_write_bed(
  linx_dir,
  sample,
  cluster,
  chainId,
  bed_outfile
)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample to examine}

\item{cluster}{Name of clusters of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be a vector of elements, where element 1 of cluster corresponds to element 1 of chainId, element 2 of cluster corresponds to element 2 of chainId ... etc}

\item{chainId}{Name of chain IDs of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be a vector of elements, where element 1 of chainId corresponds to element 1 of cluster, element 2 of chainId corresponds to element 2 of cluster ... etc}

\item{bed_outfile}{Path to output bedfile}
}
\value{
Bedfile describing ecDNA coordinates
}
\description{
Get coordinates of segments in ecDNA molecules and write to bedfile
}
