% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{read_ecdna}
\alias{read_ecdna}
\title{Read Linx ecDNA File}
\usage{
read_ecdna(linx_dir, sample, verbose = TRUE, addsample = FALSE)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample}

\item{verbose}{Verbose}

\item{addsample}{Which sample to add}
}
\value{
ecDNA (data.frame)
}
\description{
Read Linx ecDNA File
}
