% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{read_links}
\alias{read_links}
\title{Read Links File}
\usage{
read_links(linx_dir, sample, verbose = TRUE)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample}

\item{verbose}{Verbose}
}
\value{
links (data.frame)
}
\description{
Read Links File
}
