% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{read_vis_segments}
\alias{read_vis_segments}
\title{Read vis segments}
\usage{
read_vis_segments(
  linx_dir,
  sample,
  verbose = TRUE,
  fix_centromeres_and_telomeres = TRUE
)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample}

\item{verbose}{Verbose}

\item{fix_centromeres_and_telomeres}{Fix centromeres and telomeres}
}
\value{
vis segments (data.frame)
}
\description{
Read vis segments
}
