% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{read_vis_sv_data}
\alias{read_vis_sv_data}
\title{Read vis sv data}
\usage{
read_vis_sv_data(linx_dir, sample, verbose = TRUE, addsample = FALSE)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample}

\item{verbose}{Verbose}
}
\value{
vis_sv data (data.frame)
}
\description{
Read vis sv data
}
