% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{vis_segments_to_bed_df}
\alias{vis_segments_to_bed_df}
\title{vis_segments to bed}
\usage{
vis_segments_to_bed_df(vis_segments_df)
}
\arguments{
\item{vis_segments_df}{Result of \strong{read_vis_segments} (data.frame)}
}
\value{
Bed-like data.frame
}
\description{
vis_segments to bed
}
\examples{
linx_dir = system.file(package="ecDNAutils", "CHP212")
vis_segments_df <- read_vis_segments(linx_dir, "CHP212")
vis_segments_to_bed_df(vis_segments_df)
}
