% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{vis_segments_to_bed_file}
\alias{vis_segments_to_bed_file}
\title{vis_segments to bed}
\usage{
vis_segments_to_bed_file(vis_segments_df, outfile, verbose = TRUE)
}
\arguments{
\item{vis_segments_df}{result of \strong{read_vis_segments} (data.frame)}

\item{outfile}{path to output bedfile}
}
\value{

}
\description{
Writes a vis_segments dataframe to disk as a bed file
}
\examples{
linx_dir = system.file(package="utilitybeltlinx", "CHP212")
vis_segments_df <- read_vis_segments(linx_dir, "CHP212")
vis_segments_to_bed_file(vis_segments_df)
}
