% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{get_ecDNA_chain_coordinates}
\alias{get_ecDNA_chain_coordinates}
\title{Get ecDNA chain coordinates}
\usage{
get_ecDNA_chain_coordinates(linx_dir, sample, cluster = NA, chainId = NA)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files}

\item{sample}{Name of the sample to examine}

\item{cluster}{Name of clusters of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be a vector of elements, where element 1 of cluster corresponds to element 1 of chainId, element 2 of cluster corresponds to element 2 of chainId ... etc}

\item{chainId}{Name of chain IDs of interest.
By default will return all clusters/chains in ecDNA molecules.
Can be a vector of elements, where element 1 of chainId corresponds to element 1 of cluster, element 2 of chainId corresponds to element 2 of cluster ... etc}
}
\value{
Coordinates of a single ecDNA molecule (data.frame)
}
\description{
Get coordinates of segments in ecDNA molecules
}
\examples{
linx_dir = system.file(package="ecDNAutils", "CHP212")
get_ecDNA_chain_coordinates(linx_dir, "CHP212", cluster=274, chainId = 8)
}
