% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{get_ecDNA_chains}
\alias{get_ecDNA_chains}
\title{Identify ecDNA chains}
\usage{
get_ecDNA_chains(linx_dir, sample, verbose = TRUE)
}
\arguments{
\item{linx_dir}{Path to a directory containing linx files including samplename.linx.links.tsv (string)}

\item{sample}{Name of the sample of interest (string)}

\item{verbose}{Verbose (boolean)}
}
\value{
Data frame describing SV clusters and chains that comprise ecDNA molecules in a single sample
}
\description{
Lists clusters/chain combinations that comprise ecDNA molecules in a single sample
}
\examples{
linx_dir = system.file(package="ecDNAutils", "CHP212")
get_ecDNA_chains(linx_dir, "CHP212")
}
