% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_linx_data.R
\name{vis_segments_to_bed_file}
\alias{vis_segments_to_bed_file}
\title{vis_segments to bed}
\usage{
vis_segments_to_bed_file(vis_segments_df, outfile, verbose = TRUE)
}
\arguments{
\item{vis_segments_df}{Result of \strong{read_vis_segments} (data.frame)}

\item{outfile}{Path to output bedfile}
}
\value{
Run for its side effects
}
\description{
Writes a vis_segments dataframe to disk as a bed file
}
